Drag design view beyond its limits too
The entity graph view can be dragged arbitrarily so any item can be
in the center. Now we extend this to the Design view too.
